test: avoid nested conftest shadowing root fixtures
